exfat2img: fix LLVM void pointer arithmetic error
What: On LLVM and Clang with [-Werror,-Wno-pointer-arith]
void pointer arithmetic is not allowed and will cause a
compilation error.

Fix: Fix it by casting buf to (char *) before doing arithmetic.

Char * is 1 byte long, which is the same that GCC defines for void
pointer arithmetic calculations so final result is the same.

Signed-off-by: brunosampaio08 <bruno.sampaio08@unifesp.br>