fsck: fix assertion in exfat_c2o
exfat_c2o() is called with the cluster 0, and
fsck is aborted with the stack backtrace below
due to the assertion in exfat_c2o().

 exfat_c2o
 write_block
 exfat_de_iter_flush

This is caused by that write_block() could be called
for the cluster which is not read yet.

Make exfat_c2o is called only when the cluster is dirty.

Signed-off-by: Hyunchul Lee <hyc.lee@gmail.com>