exfatprogs: fix argument validation check
Since the last argument should be the device name. For the options
which required argument, if argument is not specified, the device
name will be treated as the required argument, that is unexpected.
For the options which not required argument, if the device name is
not specified, the device name pointer will be NULL, but it is not
checked.

Such as:
  # tune.exfat -L /dev/sdb1
    exfatprogs version : 1.2.4
    new label: /dev/sdb1
  # exfatlabel -i
    exfatprogs version : 1.2.4
    open failed : (null), No such file or directory

Signed-off-by: Yuezhang Mo <Yuezhang.Mo@sony.com>
Reviewed-by: Aoyama Wataru <wataru.aoyama@sony.com>