mkfs: Improve fat length computation
This accounts for fat offset and may result in smaller fat sizes.

In the optimal scenario, there are exactly as many entries in the FAT as there
are clusters on the volume + 2 unused entries in the beginning and there's no
unused space on the volume after fat offset. So we have
$$l_f = 4(n_c + 2) = \frac{l_V - o_f - n_cl_c}{n_f}$$ where

- $l_f$ - length of fat
- $l_V$ - length of volume
- $l_c$ - length of cluster
- $n_c$ - number of clusters
- $n_f$ - number of fats
- $o_f$ - fat offset

Now solving for $n_c$ we get
$$n_c = \frac{l_V - o_f - 8n_f}{l_c + 4n_f}$$

Using the $\lfloor{\frac{a - 1}{b}}\rfloor + 1$ formula to implement ceiling
division we get exactly this formula.

Signed-off-by: lvxnull <lvxnull0@protonmail.com>