mkfs: minimize initialization to zero-out on quick format mode
mkfs.exfat still zeros out the fat table and allocation bitmap even in
quick format mode. When the HDD disk grows and the cluster size is
reduced (< 4K), the fat table size becomes too large, and the time
required to zero out this can cause the format time to be excessively
long. This patch optimizes the format time by not zeroing out the fat
table.

Format time is improved from 3min to 9sec on 2TB partition with 512B
cluster size.

Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>